pyproject.toml
[tool.poetry]
name = "parser-information"
version = "0.1.0"
description = ""
authors = ["volkov_ts <volkow.timofey@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
bs4 = "^0.0.2"
chardet = "^5.2.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
django = "^5.1"
dj-database-url = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
